# packages/shared-contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "stock-research-platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/market/snapshots/{symbol}": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "List Market Snapshots",
        "operationId": "list_market_snapshots_api_v1_market_snapshots__symbol__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "title": "Symbol"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketSnapshotResponse"
                  },
                  "title": "Response List Market Snapshots Api V1 Market Snapshots  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/snapshots/{symbol}/summary": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Market Snapshot Summary",
        "operationId": "market_snapshot_summary_api_v1_market_snapshots__symbol__summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "title": "Symbol"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketSnapshotSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/bars/{symbol}": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "List Market Bars",
        "operationId": "list_market_bars_api_v1_market_bars__symbol__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "title": "Symbol"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(1m|5m|15m|30m|1h|1d)$",
              "default": "1m",
              "title": "Period"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketBarResponse"
                  },
                  "title": "Response List Market Bars Api V1 Market Bars  Symbol  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/bars/{symbol}/indicators": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "List Market Indicators",
        "operationId": "list_market_indicators_api_v1_market_bars__symbol__indicators_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "title": "Symbol"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(1m|5m|15m|30m|1h|1d)$",
              "default": "1m",
              "title": "Period"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "rsi_period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 2,
              "default": 14,
              "title": "Rsi Period"
            }
          },
          {
            "name": "macd_fast",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 2,
              "default": 12,
              "title": "Macd Fast"
            }
          },
          {
            "name": "macd_slow",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 2,
              "default": 26,
              "title": "Macd Slow"
            }
          },
          {
            "name": "macd_signal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 2,
              "default": 9,
              "title": "Macd Signal"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MarketIndicatorResponse"
                  },
                  "title": "Response List Market Indicators Api V1 Market Bars  Symbol  Indicators Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market/bars/{symbol}/cycle": {
      "get": {
        "tags": [
          "market"
        ],
        "summary": "Market Cycle Analysis",
        "operationId": "market_cycle_analysis_api_v1_market_bars__symbol__cycle_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 32,
              "title": "Symbol"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(1m|5m|15m|30m|1h|1d)$",
              "default": "1m",
              "title": "Period"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 8,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketCycleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reviews/queue": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "List Review Queue",
        "operationId": "list_review_queue_api_v1_reviews_queue_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewResponse"
                  },
                  "type": "array",
                  "title": "Response List Review Queue Api V1 Reviews Queue Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/reviews/{review_id}/decision": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Decide Review",
        "operationId": "decide_review_api_v1_reviews__review_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketBarResponse": {
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "time",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "MarketBarResponse"
      },
      "MarketCyclePeriodResponse": {
        "properties": {
          "period_bars": {
            "type": "number",
            "title": "Period Bars"
          },
          "power": {
            "type": "number",
            "title": "Power"
          }
        },
        "type": "object",
        "required": [
          "period_bars",
          "power"
        ],
        "title": "MarketCyclePeriodResponse"
      },
      "MarketCycleResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "module_version": {
            "type": "string",
            "title": "Module Version"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          },
          "fft_periods": {
            "items": {
              "$ref": "#/components/schemas/MarketCyclePeriodResponse"
            },
            "type": "array",
            "title": "Fft Periods"
          },
          "wavelet_energy": {
            "items": {
              "$ref": "#/components/schemas/MarketWaveletEnergyResponse"
            },
            "type": "array",
            "title": "Wavelet Energy"
          },
          "lstm_available": {
            "type": "boolean",
            "title": "Lstm Available"
          },
          "lstm_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lstm Reason"
          },
          "lstm_predictions": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Lstm Predictions"
          },
          "lstm_latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lstm Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "period",
          "module_version",
          "sample_count",
          "fft_periods",
          "wavelet_energy",
          "lstm_available",
          "lstm_reason",
          "lstm_predictions",
          "lstm_latency_ms"
        ],
        "title": "MarketCycleResponse"
      },
      "MarketIndicatorResponse": {
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "ma5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ma5"
          },
          "ma10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ma10"
          },
          "ma20": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ma20"
          },
          "ema5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ema5"
          },
          "ema10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ema10"
          },
          "ema20": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ema20"
          },
          "volume_ma5": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Ma5"
          },
          "volume_ma10": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Ma10"
          },
          "macd_dif": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Dif"
          },
          "macd_dea": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Dea"
          },
          "macd_hist": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Hist"
          },
          "rsi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rsi"
          }
        },
        "type": "object",
        "required": [
          "time",
          "ma5",
          "ma10",
          "ma20",
          "ema5",
          "ema10",
          "ema20",
          "volume_ma5",
          "volume_ma10",
          "macd_dif",
          "macd_dea",
          "macd_hist",
          "rsi"
        ],
        "title": "MarketIndicatorResponse"
      },
      "MarketSnapshotResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "event_time": {
            "type": "string",
            "format": "date-time",
            "title": "Event Time"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "event_time",
          "payload",
          "created_at"
        ],
        "title": "MarketSnapshotResponse"
      },
      "MarketSnapshotSummaryResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "last_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Price"
          },
          "previous_close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Close"
          },
          "change": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change"
          },
          "change_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Pct"
          },
          "bid_ask_spread": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bid Ask Spread"
          },
          "event_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Time"
          },
          "sample_count": {
            "type": "integer",
            "title": "Sample Count"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "last_price",
          "previous_close",
          "change",
          "change_pct",
          "bid_ask_spread",
          "event_time",
          "sample_count"
        ],
        "title": "MarketSnapshotSummaryResponse"
      },
      "MarketWaveletEnergyResponse": {
        "properties": {
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "energy": {
            "type": "number",
            "title": "Energy"
          }
        },
        "type": "object",
        "required": [
          "level",
          "energy"
        ],
        "title": "MarketWaveletEnergyResponse"
      },
...
      "ReviewDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "pattern": "^(APPROVED|NEEDS_REVISION|REJECTED)$",
            "title": "Decision"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "type": "object",
        "required": [
          "decision"
        ],
        "title": "ReviewDecisionRequest"
      },
      "ReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "target_type": {
            "type": "string",
            "title": "Target Type"
          },
          "target_id": {
            "type": "string",
            "title": "Target Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "target_type",
          "target_id",
          "status",
          "decision",
          "created_at"
        ],
        "title": "ReviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}